requirements.txt
# 如果 nvitop 报错"nvidia-ml-py version conflicts with other packages"
# 可以试试下面这个命令
# python -m pip install nvidia-ml-py==11.450.51

absl-py==1.3.0
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bert-score==0.3.7
blis==0.7.9
boto3==1.26.37
botocore==1.29.37
bottleneck==1.3.4
cached-path==1.1.6
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
colorlog==6.7.0
comet-ml==3.31.12
commonmark==0.9.1
confection==0.0.3
configobj==5.0.6
conllu==4.4.2
cycler==0.11.0
cymem==2.0.7
cython==0.29.32
datasets==2.4.0
defusedxml==0.7.1
deprecated==1.2.13
dill==0.3.5.1
docker-pycreds==0.4.0
dulwich==0.20.50
#en-core-web-sm==3.4.0
et-xmlfile==1.1.0
evaluate==0.3.0
everett==3.1.0
exceptiongroup==1.1.0
fairscale==0.4.6
filelock==3.7.1
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.2
gensim==3.8.3
gitdb==4.0.10
gitpython==3.1.29
google-api-core==2.11.0
google-auth-oauthlib==0.4.6
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.10.1
hydra-colorlog==1.2.0
hydra-core==1.2.0
idna==3.4
importlib-metadata==5.1.0
info-nce-pytorch==0.1.4
iniconfig==1.1.1
jellyfish==0.9.0
jieba==0.42.1
jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonlines==3.1.0
jsonnet==0.19.1
jsonschema==4.17.3
keybert==0.6.0
kiwisolver==1.4.2
langcodes==3.3.0
lmdb==1.4.0
lxml==4.9.1
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==8.14.0
multidict==6.0.3
multiprocess==0.70.13
munkres==1.1.4
murmurhash==1.0.9
networkx==2.8.8
nlg-eval==2.3
nltk==3.7
numexpr==2.8.1
numpy==1.24.0
nvidia-ml-py==11.450.51
nvitop==0.8.0
oauthlib==3.2.2
omegaconf==2.2.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
pathy==0.10.1
pillow==9.0.1
pip==22.3.1
pluggy==1.0.0
portalocker==2.6.0
preshed==3.0.8
promise==2.3
protobuf==3.19.6
psutil==5.9.2
pyarrow==10.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.8.2
pydeprecate==0.3.2
pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
pytorch-lightning==1.7.5
pytz==2022.1
pyyaml==6.0
rake-nltk==1.0.6
redis==4.3.4
regex==2022.10.31
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests==2.28.1
responses==0.18.0
rich==12.5.1
rouge-score==0.1.2
rouge==1.0.1
rsa==4.9
s3transfer==0.6.0
sacrebleu==2.2.0
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.3
seaborn==0.12.0
segtok==1.5.11
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
setuptools==65.5.0
shortuuid==1.0.11
shyaml==0.6.2
#simcse==0.4
#sip==4.19.13
six==1.16.0
smart-open==6.2.0
smmap==5.0.0
spacy-legacy==3.0.10
spacy-loggers==1.0.4
spacy==3.3.2
srsly==2.4.5
stanfordcorenlp==3.9.1.1
tabulate==0.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.0
tensorboardx==2.5.1
termcolor==1.1.0
theano==1.0.5
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.10.3
tomli==2.0.1
torchmetrics==0.11.0
tornado==6.1
tqdm==4.64.1
traitlets==5.8.0
transformers==4.15.0
typer==0.4.2
typing-extensions==4.4.0
urllib3==1.26.13
wandb==0.12.21
wasabi==0.10.1
wcwidth==0.2.5
websocket-client==1.3.3
werkzeug==2.2.2
wheel==0.37.1
word2number==1.1
wrapt==1.14.1
wurlitzer==3.0.3
xdg==5.1.1
xxhash==3.1.0
yake==0.4.8
yarl==1.8.2
zipp==3.11.0

allennlp-models==2.10.1
allennlp==2.10.1
py-rouge==1.1  # for q2